Working Hex instantiation in layers
Also gives each hex its own coordinates and have prep for adding additional sprites and switching between them (may have to end up being different prefabs for each star-sign, not sure)

+++ Game/ConstTileAtion/Assets/Scripts/HexInfo.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class HexInfo : MonoBehaviour {
 
     //Info for each hex
-    public float X, Y, Z;
+    public int X, Y, Z;
     public int Layer;
     public GameObject HexGenerator;
